+++ system_tests/test_id_token.py
@@ -1,4 +1,4 @@
-# Copyright 2016 Google LLC
+# Copyright 2020 Google LLC
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -11,16 +11,15 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+import pytest
 
-"""Google Auth Library for Python."""
-
-import logging
-
-from google.auth._default import default, load_credentials_from_file
+from google.auth import jwt
+import google.oauth2.id_token
 
 
-__all__ = ["default", "load_credentials_from_file"]
+def test_fetch_id_token(http_request):
+    audience = "https://pubsub.googleapis.com"
+    token = google.oauth2.id_token.fetch_id_token(http_request, audience)
 
-
-# Set default logging handler to avoid "No handler found" warnings.
-logging.getLogger(__name__).addHandler(logging.NullHandler())
+    _, payload, _, _ = jwt._unverified_decode(token)
+    assert payload["aud"] == audience